+++ x/pocketcore/types/evidence_test.go
@@ -31,6 +31,7 @@ func TestAllEvidence_AddGetEvidence(t *testing.T) {
 	}
 	proof := RelayProof{
 		Entropy:            0,
+		RequestHash:        header.HashString(), // fake
 		SessionBlockHeight: 1,
 		ServicerPubKey:     servicerPubKey,
 		Blockchain:         ethereum,
@@ -46,7 +47,7 @@ func TestAllEvidence_AddGetEvidence(t *testing.T) {
 	if err != nil {
 		t.Fatalf(err.Error())
 	}
-	assert.True(t, reflect.DeepEqual(GetEvidenceMap().GetProof(header, 0), proof))
+	assert.True(t, reflect.DeepEqual(GetEvidenceMap().GetProof(header, RelayEvidence, 0), proof))
 }
 
 func TestAllEvidence_DeleteEvidence(t *testing.T) {
@@ -72,6 +73,7 @@ func TestAllEvidence_DeleteEvidence(t *testing.T) {
 		Entropy:            0,
 		SessionBlockHeight: 1,
 		ServicerPubKey:     servicerPubKey,
+		RequestHash:        header.HashString(), // fake
 		Blockchain:         ethereum,
 		Token: AAT{
 			Version:              "0.0.1",
@@ -85,10 +87,10 @@ func TestAllEvidence_DeleteEvidence(t *testing.T) {
 	if err != nil {
 		t.Fatalf(err.Error())
 	}
-	assert.True(t, reflect.DeepEqual(GetEvidenceMap().GetProof(header, 0), proof))
-	GetEvidenceMap().GetProof(header, 0)
-	GetEvidenceMap().DeleteEvidence(header)
-	assert.Empty(t, GetEvidenceMap().GetProof(header, 0))
+	assert.True(t, reflect.DeepEqual(GetEvidenceMap().GetProof(header, RelayEvidence, 0), proof))
+	GetEvidenceMap().GetProof(header, RelayEvidence, 0)
+	GetEvidenceMap().DeleteEvidence(header, RelayEvidence)
+	assert.Empty(t, GetEvidenceMap().GetProof(header, RelayEvidence, 0))
 }
 
 func TestAllEvidence_GetProofs(t *testing.T) {
@@ -113,6 +115,7 @@ func TestAllEvidence_GetProofs(t *testing.T) {
 	proof := RelayProof{
 		Entropy:            0,
 		SessionBlockHeight: 1,
+		RequestHash:        header.HashString(), // fake
 		ServicerPubKey:     servicerPubKey,
 		Blockchain:         ethereum,
 		Token: AAT{
@@ -126,6 +129,7 @@ func TestAllEvidence_GetProofs(t *testing.T) {
 	proof2 := RelayProof{
 		Entropy:            1, // just for testing equality
 		SessionBlockHeight: 1,
+		RequestHash:        header.HashString(), // fake
 		ServicerPubKey:     servicerPubKey,
 		Blockchain:         ethereum,
 		Token: AAT{
@@ -141,7 +145,7 @@ func TestAllEvidence_GetProofs(t *testing.T) {
 	if err != nil {
 		t.Fatalf(err.Error())
 	}
-	proofs := GetEvidenceMap().GetProofs(header)
+	proofs := GetEvidenceMap().GetProofs(header, RelayEvidence)
 	assert.NotNil(t, proofs)
 	assert.Len(t, proofs, 2)
 	assert.Equal(t, proofs[0], proof)
@@ -176,6 +180,7 @@ func TestAllEvidence_GetTotalRelays(t *testing.T) {
 		Entropy:            0,
 		SessionBlockHeight: 1,
 		ServicerPubKey:     servicerPubKey,
+		RequestHash:        header.HashString(), // fake
 		Blockchain:         ethereum,
 		Token: AAT{
 			Version:              "0.0.1",
@@ -189,6 +194,7 @@ func TestAllEvidence_GetTotalRelays(t *testing.T) {
 		Entropy:            0,
 		SessionBlockHeight: 1,
 		ServicerPubKey:     servicerPubKey,
+		RequestHash:        header.HashString(), // fake
 		Blockchain:         ethereum,
 		Token: AAT{
 			Version:              "0.0.1",