+++ x/pocketcore/types/proof.go
@@ -9,15 +9,62 @@ import (
 )
 
 type Proof interface {
-	Validate(maxRelays int64, numberOfChains, sessionNodeCount int, sessionBlockHeight int64, hb HostedBlockchains, verifyPubKey string) sdk.Error
+	Validate(maxRelays int64, numberOfChains, sessionNodeCount int, sessionBlockHeight int64, hb HostedBlockchains, payloadHash, verifyPubKey string) sdk.Error
 	Hash() []byte
 	HashString() string
 	HashWithSignature() []byte
 	HashStringWithSignature() string
 }
 
+type ChallengeProofInvalidData struct {
+}
+
+func (c ChallengeProofInvalidData) Validate(maxRelays int64, numberOfChains, sessionNodeCount int, sessionBlockHeight int64, hb HostedBlockchains, payloadHash, verifyPubKey string) sdk.Error {
+	panic("implement me")
+}
+
+func (c ChallengeProofInvalidData) Hash() []byte {
+	panic("implement me")
+}
+
+func (c ChallengeProofInvalidData) HashString() string {
+	panic("implement me")
+}
+
+func (c ChallengeProofInvalidData) HashWithSignature() []byte {
+	panic("implement me")
+}
+
+func (c ChallengeProofInvalidData) HashStringWithSignature() string {
+	panic("implement me")
+}
+
+type ChallengeProofCorruptedRequest struct {
+}
+
+func (c ChallengeProofCorruptedRequest) Validate(maxRelays int64, numberOfChains, sessionNodeCount int, sessionBlockHeight int64, hb HostedBlockchains, payloadHash, verifyPubKey string) sdk.Error {
+	panic("implement me")
+}
+
+func (c ChallengeProofCorruptedRequest) Hash() []byte {
+	panic("implement me")
+}
+
+func (c ChallengeProofCorruptedRequest) HashString() string {
+	panic("implement me")
+}
+
+func (c ChallengeProofCorruptedRequest) HashWithSignature() []byte {
+	panic("implement me")
+}
+
+func (c ChallengeProofCorruptedRequest) HashStringWithSignature() string {
+	panic("implement me")
+}
+
 // RelayProof per relay
 type RelayProof struct {
+	RequestHash        string `json:"request_hash"`
 	Entropy            int64  `json:"entropy"`
 	SessionBlockHeight int64  `json:"session_block_height"`
 	ServicerPubKey     string `json:"servicer_pub_key"`
@@ -26,38 +73,42 @@ type RelayProof struct {
 	Signature          string `json:"signature"`
 }
 
-func (p RelayProof) Validate(maxRelays int64, numberOfChains, sessionNodeCount int, sessionBlockHeight int64, hb HostedBlockchains, verifyPubKey string) sdk.Error {
+func (rp RelayProof) Validate(maxRelays int64, numberOfChains, sessionNodeCount int, sessionBlockHeight int64, hb HostedBlockchains, payloadHash, verifyPubKey string) sdk.Error {
 	// validate the session block height
-	if p.SessionBlockHeight != sessionBlockHeight {
+	if rp.SessionBlockHeight != sessionBlockHeight {
 		return NewInvalidBlockHeightError(ModuleName)
 	}
 	// validate blockchain
-	if err := HashVerification(p.Blockchain); err != nil {
+	if err := HashVerification(rp.Blockchain); err != nil {
 		return err
 	}
+	// validate request hash
+	if payloadHash != rp.RequestHash {
+		return NewRequestHashError(ModuleName)
+	}
 	evidenceHeader := SessionHeader{
-		ApplicationPubKey:  p.Token.ApplicationPublicKey,
-		Chain:              p.Blockchain,
-		SessionBlockHeight: p.SessionBlockHeight,
+		ApplicationPubKey:  rp.Token.ApplicationPublicKey,
+		Chain:              rp.Blockchain,
+		SessionBlockHeight: rp.SessionBlockHeight,
 	}
 	// validate not over service
 	totalRelays := GetEvidenceMap().GetTotalRelays(evidenceHeader)
-	if !GetEvidenceMap().IsUniqueProof(evidenceHeader, p) {
+	if !GetEvidenceMap().IsUniqueProof(evidenceHeader, rp) {
 		return NewDuplicateProofError(ModuleName)
 	}
 	if totalRelays >= int64(math.Ceil(float64(maxRelays)/float64(numberOfChains))/(float64(sessionNodeCount))) {
 		return NewOverServiceError(ModuleName)
 	}
 	// validate the public key correctness
-	if p.ServicerPubKey != verifyPubKey {
+	if rp.ServicerPubKey != verifyPubKey {
 		return NewInvalidNodePubKeyError(ModuleName) // the public key is not this nodes, so they would not get paid
 	}
 	// ensure the blockchain is supported
-	if !hb.ContainsFromString(p.Blockchain) {
+	if !hb.ContainsFromString(rp.Blockchain) {
 		return NewUnsupportedBlockchainNodeError(ModuleName)
 	}
 	// validate the RelayProof public key format
-	if err := PubKeyVerification(p.ServicerPubKey); err != nil {
+	if err := PubKeyVerification(rp.ServicerPubKey); err != nil {
 		return NewInvalidNodePubKeyError(ModuleName)
 	}
 	// validate the verify public key format
@@ -65,31 +116,33 @@ func (p RelayProof) Validate(maxRelays int64, numberOfChains, sessionNodeCount i
 		return NewInvalidNodePubKeyError(ModuleName)
 	}
 	// validate the service token
-	if err := p.Token.Validate(); err != nil {
+	if err := rp.Token.Validate(); err != nil {
 		return NewInvalidTokenError(ModuleName, err)
 	}
-	return SignatureVerification(p.Token.ClientPublicKey, p.HashString(), p.Signature)
+	return SignatureVerification(rp.Token.ClientPublicKey, rp.HashString(), rp.Signature)
 }
 
 // structure used to json marshal the RelayProof
-type relayRelayProof struct {
+type relayProof struct {
 	Entropy            int64  `json:"entropy"`
 	SessionBlockHeight int64  `json:"session_block_height"`
 	ServicerPubKey     string `json:"servicer_pub_key"`
 	Blockchain         string `json:"blockchain"`
 	Signature          string `json:"signature"`
 	Token              string `json:"token"`
+	RequestHash        string `json:"request_hash"`
 }
 
 // convert the RelayProof to bytes
-func (p RelayProof) Bytes() []byte {
-	res, err := json.Marshal(relayRelayProof{
-		Entropy:            p.Entropy,
-		ServicerPubKey:     p.ServicerPubKey,
-		Blockchain:         p.Blockchain,
-		SessionBlockHeight: p.SessionBlockHeight,
+func (rp RelayProof) Bytes() []byte {
+	res, err := json.Marshal(relayProof{
+		Entropy:            rp.Entropy,
+		RequestHash:        rp.RequestHash,
+		ServicerPubKey:     rp.ServicerPubKey,
+		Blockchain:         rp.Blockchain,
+		SessionBlockHeight: rp.SessionBlockHeight,
 		Signature:          "", // omit the signature
-		Token:              p.Token.HashString(),
+		Token:              rp.Token.HashString(),
 	})
 	if err != nil {
 		panic(fmt.Sprintf("an error occured converting the relay RelayProof to bytes:\n%v", err))
@@ -98,14 +151,15 @@ func (p RelayProof) Bytes() []byte {
 }
 
 // convert the RelayProof to bytes
-func (p RelayProof) BytesWithSignature() []byte {
-	res, err := json.Marshal(relayRelayProof{
-		Entropy:            p.Entropy,
-		ServicerPubKey:     p.ServicerPubKey,
-		Blockchain:         p.Blockchain,
-		SessionBlockHeight: p.SessionBlockHeight,
-		Signature:          p.Signature,
-		Token:              p.Token.HashString(),
+func (rp RelayProof) BytesWithSignature() []byte {
+	res, err := json.Marshal(relayProof{
+		Entropy:            rp.Entropy,
+		ServicerPubKey:     rp.ServicerPubKey,
+		RequestHash:        rp.RequestHash,
+		Blockchain:         rp.Blockchain,
+		SessionBlockHeight: rp.SessionBlockHeight,
+		Signature:          rp.Signature,
+		Token:              rp.Token.HashString(),
 	})
 	if err != nil {
 		panic(fmt.Sprintf("an error occured converting the relay RelayProof to bytes with signature:\n%v", err))
@@ -114,23 +168,23 @@ func (p RelayProof) BytesWithSignature() []byte {
 }
 
 // addr the RelayProof bytes
-func (p RelayProof) Hash() []byte {
-	res := p.Bytes()
+func (rp RelayProof) Hash() []byte {
+	res := rp.Bytes()
 	return Hash(res)
 }
 
 // hex encode the RelayProof addr
-func (p RelayProof) HashString() string {
-	return hex.EncodeToString(p.Hash())
+func (rp RelayProof) HashString() string {
+	return hex.EncodeToString(rp.Hash())
 }
 
 // addr the RelayProof bytes
-func (p RelayProof) HashWithSignature() []byte {
-	res := p.BytesWithSignature()
+func (rp RelayProof) HashWithSignature() []byte {
+	res := rp.BytesWithSignature()
 	return Hash(res)
 }
 
 // hex encode the RelayProof addr
-func (p RelayProof) HashStringWithSignature() string {
-	return hex.EncodeToString(p.HashWithSignature())
+func (rp RelayProof) HashStringWithSignature() string {
+	return hex.EncodeToString(rp.HashWithSignature())
 }