+++ doc/changelog.md
@@ -7,3 +7,4 @@
 - Update posmint module to use sdk.Ctx interface
 - Fix `pseudorandomGenerator` unexported properties would return empty json
 - Evidence now holds proof interface to allow for challenge proofs 
+- Added Relay Request Hash to RelayProof object